contracts-stylus: contracts: core: Split up core contract
