+++ contracts-stylus/src/contracts/test_contracts/darkpool_test_contract.rs
@@ -7,10 +7,20 @@ use contracts_common::constants::{
     DARKPOOL_CORE_ADDRESS_SELECTOR, MERKLE_ADDRESS_SELECTOR, TRANSFER_EXECUTOR_ADDRESS_SELECTOR,
     VERIFIER_ADDRESS_SELECTOR, VKEYS_ADDRESS_SELECTOR,
 };
-use stylus_sdk::{alloy_primitives::U256, prelude::*};
+use stylus_sdk::{
+    alloy_primitives::{Address, U256},
+    prelude::*,
+    storage::{StorageArray, StorageBool, StorageMap, StorageU256},
+};
 
 use crate::{
-    contracts::{darkpool::DarkpoolContract, darkpool_core::DarkpoolCoreContract},
+    contracts::{
+        core::{
+            core_helpers::mark_nullifier_spent, core_wallet_ops::CoreWalletOpsContract,
+            CoreContractStorage,
+        },
+        darkpool::DarkpoolContract,
+    },
     utils::{
         helpers::{delegate_call_helper, u256_to_scalar},
         solidity::{init_0Call as initMerkleCall, isDummyUpgradeTargetCall},
@@ -26,6 +36,44 @@ struct DarkpoolTestContract {
     darkpool: DarkpoolContract,
 }
 
+impl CoreContractStorage for DarkpoolTestContract {
+    fn verifier_address(&self) -> Address {
+        self.darkpool.verifier_address.get()
+    }
+
+    fn vkeys_address(&self) -> Address {
+        self.darkpool.vkeys_address.get()
+    }
+
+    fn merkle_address(&self) -> Address {
+        self.darkpool.merkle_address.get()
+    }
+
+    fn transfer_executor_address(&self) -> Address {
+        self.darkpool.transfer_executor_address.get()
+    }
+
+    fn nullifier_set(&self) -> &StorageMap<U256, StorageBool> {
+        &self.darkpool.nullifier_set
+    }
+
+    fn nullifier_set_mut(&mut self) -> &mut StorageMap<U256, StorageBool> {
+        &mut self.darkpool.nullifier_set
+    }
+
+    fn public_blinder_set(&self) -> &StorageMap<U256, StorageBool> {
+        &self.darkpool.public_blinder_set
+    }
+
+    fn public_blinder_set_mut(&mut self) -> &mut StorageMap<U256, StorageBool> {
+        &mut self.darkpool.public_blinder_set
+    }
+
+    fn protocol_public_encryption_key(&self) -> &StorageArray<StorageU256, 2> {
+        &self.darkpool.protocol_public_encryption_key
+    }
+}
+
 // We manually implement `Borrow` and `BorrowMut` to enable the `DarkpoolTestContract` to
 // call the internal methods of the `DarkpoolCoreContract` on the nested `DarkpoolContract`.
 // We do this by unsafely casting a pointer to the `DarkpoolContract` to a pointer to the
@@ -34,23 +82,23 @@ struct DarkpoolTestContract {
 // This is possible because we already maintain that the `DarkpoolContract` and
 // `DarkpoolCoreContract` have exactly the same storage / memory layout.
 
-impl Borrow<DarkpoolCoreContract> for DarkpoolTestContract {
-    fn borrow(&self) -> &DarkpoolCoreContract {
+impl Borrow<CoreWalletOpsContract> for DarkpoolTestContract {
+    fn borrow(&self) -> &CoreWalletOpsContract {
         unsafe {
             let darkpool_ptr: *const DarkpoolContract = &self.darkpool;
-            let darkpool_core_ptr: *const DarkpoolCoreContract =
-                darkpool_ptr as *const DarkpoolCoreContract;
+            let darkpool_core_ptr: *const CoreWalletOpsContract =
+                darkpool_ptr as *const CoreWalletOpsContract;
             &*darkpool_core_ptr
         }
     }
 }
 
-impl BorrowMut<DarkpoolCoreContract> for DarkpoolTestContract {
-    fn borrow_mut(&mut self) -> &mut DarkpoolCoreContract {
+impl BorrowMut<CoreWalletOpsContract> for DarkpoolTestContract {
+    fn borrow_mut(&mut self) -> &mut CoreWalletOpsContract {
         unsafe {
             let darkpool_ptr: *mut DarkpoolContract = &mut self.darkpool;
-            let darkpool_core_ptr: *mut DarkpoolCoreContract =
-                darkpool_ptr as *mut DarkpoolCoreContract;
+            let darkpool_core_ptr: *mut CoreWalletOpsContract =
+                darkpool_ptr as *mut CoreWalletOpsContract;
             &mut *darkpool_core_ptr
         }
     }
@@ -62,7 +110,7 @@ impl DarkpoolTestContract {
     /// Marks the given nullifier as spent
     pub fn mark_nullifier_spent(&mut self, nullifier: U256) -> Result<(), Vec<u8>> {
         let nullifier = u256_to_scalar(nullifier)?;
-        DarkpoolCoreContract::mark_nullifier_spent(self, nullifier)
+        mark_nullifier_spent::<Self, _>(self, nullifier)
     }
 
     /// Attempts to call [`DummyUpgradeTarget::is_dummy_upgrade_target`] on either
